Create  Cube
ime limit per test1 second
memory limit per test256 megabytes
In geometry, a cube is a three-dimensional solid object bounded by six square faces, with three meeting at each vertex. The image bellow is an example of a cube.


You are given the surface area of a cube, and your task is to find the length of that cube's edge. Can you?

Input
The first line contains an integer T (1 ≤ T ≤ 1000), in which T is the number of test cases.

Each test case consists of a line containing an integer a (1 ≤ a ≤ 6 × 106), giving the surface area of a cube.

Output
For each test case, print a single line containing the length of the given cube's edge.

It is guaranteed that all answers are integer numbers. Do not print any floating-point values.

Example
InputCopy
2
24
6
OutputCopy
2
1

*******************
they give me the area of the cube and it has 6 side every one is equal the another then i devide the area by 6 the i need to take the sqrt to find the length